% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialized-package.R
\docType{package}
\name{geospatialized-package}
\alias{geospatialized}
\alias{geospatialized-package}
\title{geospatialized: What the Package Does (One Line, Title Case)}
\description{
Geospatial analysis functions for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/geospatialized}
  \item Report bugs at \url{https://github.com/johnaclouse/geospatialized/issues}
}

}
\author{
\strong{Maintainer}: John Clouse \email{john@clouse.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
