% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_geometry.R
\name{get_acs_geometry}
\alias{get_acs_geometry}
\title{Title}
\usage{
get_acs_geometry(states_needed = "Oregon")
}
\arguments{
\item{states_needed}{character vector of state geometries to be returned}
}
\value{
a list of lists containing shape geometry and total population at the
state, county, census tract, and block group level.
}
\description{
Title
}
\examples{
get_acs_geometry()
}
