% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_variable_catalog.R
\name{get_acs_variable_catalog}
\alias{get_acs_variable_catalog}
\title{get_acs_variable_catalog}
\usage{
get_acs_variable_catalog(
  year = 2019,
  dataset = "acs5",
  label_filter = "*",
  concept_filter = "*"
)
}
\arguments{
\item{year}{numeric effective year for ACS data}

\item{dataset}{character typcially set to "acs5"}

\item{label_filter}{character use "*" to accept all variables or a character
to filter the returned variabls}

\item{concept_filter}{character use "*" to accept all concepts or a character
value to filter the returned concepts}
}
\value{
tibble containing name, label, and concept for 27,000 + ACS fields
}
\description{
get_acs_variable_catalog
}
\examples{
\dontrun{
View(get_acs_variable_catalog())
View(get_acs_variable_catalog(label_filter = "transportation", concept_filter = "vehicle"))
}
}
