% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_states_fips_codes.R
\name{get_state_fips_codes}
\alias{get_state_fips_codes}
\title{get_state_fips_codes}
\usage{
get_state_fips_codes(states)
}
\arguments{
\item{states}{character vector name of states to lookup

Returns the FIPS codes for the requested states.}
}
\value{
character vector with numeric fips codes
}
\description{
get_state_fips_codes
}
\examples{
get_state_fips_codes("Oregon")
}
