% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOTO_M1}
\alias{GOTO_M1}
\title{geospatialized test location}
\format{
A tibble
}
\usage{
GOTO_M1
}
\description{
Location of the original HC Kendall GOTO Mars Model M-1 planetarium star projector.
https://www.goto.co.jp/english/work/transition/
}
\keyword{datasets}
