% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geographic_areas.R
\name{plot_geographic_areas}
\alias{plot_geographic_areas}
\title{plot_geographic_areas}
\usage{
plot_geographic_areas(
  location = geospatialized::GOTO_M1,
  states_needed = "Oregon",
  acs_geometry = NULL
)
}
\arguments{
\item{location}{data frame with columns "lon" and "lat"}

\item{states_needed}{character vector containing states needed for geography geometeries}

\item{acs_geometry}{a list of acs_geometries from \code{get_acs_geometry()}}
}
\value{
leaflet object
}
\description{
Plot the state, county, census tract, and block group geographic areas containing the target_location.
}
