% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoid.R
\name{get_geoid}
\alias{get_geoid}
\title{get_geoid}
\usage{
get_geoid(lon, lat, acs_geometry = get_acs_geometry()$block_groups)
}
\arguments{
\item{lon}{numeric longitude of target location}

\item{lat}{numeric latitude of target location}

\item{acs_geometry}{list of acs geometries

If the acs_geometry is at the block group level, the returned GEOID will be
the block group of the target location. Likewise, if the acs_geometry is at
the census tract level, the GEOID returned will be the census tract
containing the target location.}
}
\value{
charcter value of GEOID (FIPS code)
}
\description{
get_geoid
}
\examples{
\dontrun{
get_geoid(lon = GOTO_M1$lon, lat = GOTO_M1$lat)
}
}
